backend/requirements.txt
fastapi>=0.100.0
python-dotenv>=1.0.0
livekit-api>=1.0.7
uvicorn>=0.30.0
